fixes: pwm and others
- (somewhat) fix build script for windows
- fix clang-format (and run it, I forgot last commit oops)
- make some sort of porting guide
- add more default functions in int.h
- change pwm_read_raw to read a pulsewifth instead of a duty cycle
- fix both pwm read and write for pico (esp32 soon[tm])
- document "example" port
- fix receiver, servo, and esc pwm implementations (they were all wrong)
- fix some memory management compiler warnings
- update log system to dynamically allocate memory

+++ platform/example/pwm.c
@@ -6,13 +6,24 @@
 #include "platform/pwm.h"
 
 bool pwm_setup_read(u32 pins[], u32 num_pins) {
+    // This function will be called before executing any other PWM READ-related functions on a given pin.
+    // It should run any prerequisite setup for reading PWM signals on the given pins.
+    // It should return true if the setup was successful, false if not.
 }
 
 bool pwm_setup_write(u32 pins[], u32 num_pins, u32 freq) {
+    // This function will be called before executing any other PWM WRITE-related functions on a given pin.
+    // It should run any prerequisite setup for writing PWM signals on the given pins at the given frequency (in Hz).
+    // It should return true if the setup was successful, false if not.
 }
 
-i32 pwm_read_raw(u32 pin) {
+float pwm_read_raw(u32 pin) {
+    // This function should return the pulsewidth cycle of the PWM signal on the given pin.
+    // The duty cycle should be returned in the form of a floating-point number, measured in μs (microseconds).
+    // If the pin is invalid or some sort of error occurs, return -1.
 }
 
 void pwm_write_raw(u32 pin, u16 duty) {
+    // This function should write the given duty cycle to the given pin.
+    // The duty cycle is an unsigned 16-bit integer, so it ranges from 0 to 2^16 [65535] (0 being off and 2^16 being on).
 }
